Add files via upload
Added  I2C_2004_LCD driver for BBC Microbit/MicroPython. This is a modified version of Shaoziyang's I2C_1602_LCD driver. Have also written two demo programs to show how it is used with the Microbit. Also included is an I2C Address Scanner program for the Microbit and several photos to help illustrate issues discussed in the ReadMe.txt file.